♻️Refactor tests, adding a testConfig file & timeouts
